// frontend/src/api/generated/models/pettyCashResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface PettyCashResponse {
  id: string;
  date: string;
  type: string;
  amount: string;
  category: string | null;
  description: string | null;
  status: string;
  pic: string | null;
}